ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
xgboost==2.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
